domains via conditions and iterations in lambda parameters
Domains are now internally represented as conditions, just with
different shorthand notation.  This simplifies a lot of things (don't
need different axioms/theorems that depend upon whether there is a
domain or not).  Allow lambda parameters, including "forall" instance
variables, to be iterations that may be relabelled (or directly
specialized in the "forall" case).