implementing "styles"
Also added InnerExpr class and Expression.innerExpr() method to generate
it.  This is useful for changing the style of an inner expression within
an expression (using a "with..." method call) or create a lambda
expression for changing an inner expression.   The style does not affect
the meaning of an expression.  Changing the style of an expression will
change it for any Expression, KnownTruth, or Proof containing it.