Allow ability to manually disable proofs and fixing way proofs update
When a Proof is disabled, we should find the best remaining Proof for a
KnownTruth.  This was not done properly before and could miss
possibilities.  In the process of correcting this, we change the way
internal information of  Prove-It objects (Expression, KnownTruth, and
Proof) is stored, using newly defined_MeaningData and _StyleData objects
for storing internal information in a less redundant manner than before.