Adds an allConditionsList() method to operation_over_instances.py
Adds an allConditionsList() method to operation_over_instances.py, which returns an actual list of the conditions rather than the generator obtained using allConditions(). The allConditionsList() method uses the allConditions() generator, though, so the allConditions() method has been retained in the operation_over_instances.py code. This commit also updates the corresponding cell in the tutorial05_forall notebook which was trying to use an outdated conditions attribute.