Fix typos in generator fxns & augment with list fxns
Fixes mis-calls or typos in several generator functions in
operation_over_instances.py and augments the code with fxns that return
actual lists instead of generators (in the allInstanceVars() and allDomains()
methods). Updates introductory portion of tutorial05_forall notebook to
be consistent with operation_over_instances.py fixes, to update and fix some
comments, and to add a more complex Forall() example to help illustrate
some of the attributes and methods. Tutorial notebook now works well up to the Specialization section.