Update tutorial05_forall to full functionality.
tutorial05_forall now fully functional with all input cells executing
successfully, including now some minor changes in the section on
generalization. A key change in that section was the creation of a more
appropriate specialization of the substitution known truth, which then reacted
more predictably (or more as expected) to the generalize() attempts.