Crude relabel() method for both sides of turnstile
Achieves a crude version of a relabel() method, called relabelFully(), in the
KnownTruth class, that achieves (under limited circumstances) a relabeling
of variables on both sides of a KnownTruth turnstile operator. This
temporarily uses a new class Specialize_02 based on the Specialize class
found in proveit._core_.proof.py. NOT at all ready for anyone else to see
or actively utilize.