Return instead of yield in the Exists.sideEffect() generator
Temporary modification to the Exists.sideEffect() method to simply return
rather than yield, to avoid side-effect-based errors and allow experimenting
to continue. Includes a quick commit of some changes in the epxloration
jupyter notebook and the deletion of an old notebook that was saving some
old (error-related) work (the Specialize_Forall_Excerpt_Error.ipynb).