Establish usedLiterals() in Expression and Literal classes
First commit that establishes the usedLiterals() method in the Expression
and Literal classes, in preparation for a skolemize() derivation method.
We eventually want to track all ''current'' Literals so that a skolemize()
method can check to ensure we are not re-using a Literal already being
used as a concrete instantiation. This commit also includes some changes
in the skolemization_exploration notebook, reflecting just some
exprimenting with the code and newly-introduced methods.