Introduced temporary skolemize2() function
The skolemize2() function appears temporarily in the
skolemization_explorations.ipynb notebook, and provides the ability
the skolemize any one or more of a collection of nested Exists
operations. For example, one could skolemize all three of a triplet of
nested Exists, or one could skolemize just the middle of a triplet of
nested Exists. Not yet clear how to incorporate this into the simpler
(but already complex) KnownTruth skolemize() method.