Merge Sqrt(2) Demo into skolemization branch
Merged an old Sqrt(2) branch into this craft_edits_skolemization
branch, allowing us to continue working on details of the "Sqrt(2)
not in Q" proof. The merge seemed to have worked fine, with only
one easier-to-manage conflict (in the exists.py file where I needed
to work around a sideeffects-yielding function). This commit also
deletes a number of print statements in the proof.py and
known_truth.py that were being used to understand the
specialize() method and the Specialization class.