Further steps in the implementation of Sqrt(2) Proof
Includes a shift of numerous assumptions into the _theorems_ of
proveit/number/sets (meant to be a temporary location) and a
related reorganizing of the number/sets/_demonstrations_.ipynb
notebook where we are housing the sqrt(2) not in Rationals
proof efforts. Also includes separate GCD function class definition
and separate Divides operation class definition.