"Fraction" -> "frac", "Frac" -> "frac", and "Subtract" -> "subtract"
Since "subtract" is now a method that generates an Add expression with a subtraction style, use lowercase to be consistent with the convention that functions are lowercase.  Similarly, "Frac" (which used to be "Fraction") should be "frac" since it is a method that generates a Div expression.  This also serves as a reminder/indicator that there is no Frac or Subract expression (anymore) -- they are just stylized versions of other kinds of expressions.