Example for issue #28.
Example that motivates making Prove-It search for a connection between KnownTruths as a last resort to see of all of the pieces to prove something are already "known".  The idea will be to raise an exception that indicates to the user that all of the pieces are there but need to be put together properly for an efficient proof.