Further work on sqrt(2) not rational proof
Adds a skolemElim theorem to the theorems for the sqrt(2) proof and makes
a little progress on applying it to the elimination of assumptions at the end
of the proof. Also reworks a few items in the proof-building steps to convert
assumptions to tuples of items rather than logical ands.