Begin rework of membership deduceInBool() implementation
After text/discussion with Wayne, realized was approaching this wrong.
See comments in Issue #59 . Eliminated the membershipInBool axiom and
related notInSetInBool theorem, but added a new same-named theorem to
indicate that (x in S) in Bool gives (x notIn S) in Bool. Still to be dealt with:
implementing the correct deduceInBool() for the InSet and NotInSet classes.