Add three xInYInBool theorems to number/sets/integer
Related to Issue #59, the theorems added to proveit/number/sets/integer
are: xInNatsInBool, xInNatsPosInBool, xInIntsInBool. None of these are proven
yet. They correspond to the sets Naturals, NaturalsPos, and Integers defined
in the sets/integer/_axioms_ and the corresponding NumberSets defined in
the python code in number/sets/naturals.py and number/sets/integers.py.