Add deduceInBool() method to NumberMembership class.
Added a deduceInBool() method to the NumberMembership class (in number/sets/number_set.py) and a deduceMembershipInBool() method in the NaturalsSet class (in number/sets/integer/naturals.py), the combination of which now seems to work to allow InSet(n, Naturals) to produce |– (n \in N) \in Bool. Hoping I can use this now as a template for the other sets. Related to Issue #59.