Continue establishing deduceInBool() methods
Finished establishing the xInYInBool theorems for the rest of the
currently-defined NumberSets (now includes thms for RealsPos and RealsNeg), and established the deduceInBool() methods for the remaining NumberSet classes of NaturalsPosSet, IntegerSet, RealsSet, RealsPosSet,
and RealsNegSet. Brief testing in demonstrations pages confirmed that
InSet(x, S).deduceInBool() works now for each of these predefined sets. Also began introductory edits of the corresponding demonstration nbs for the integer, real, and complex NumberSets. Issue #59 .