Establish min working version of Or.concludeViaPermutation
Related Issues: #69 , #75 , #77 .
Commit also includes a temporary _test_ ipynb for working on the
Or.concludeViaSome and Or.concludeViaPermutation methods. The
concludeViaPermutation is a toy version that generates an arbitrary
permutation of the permuted_disjunction (putting the 1st operand at
the end of the disjunction), without actually trying to commute everything
to match the ultimately-desired order of operands. This took longer than
I had hoped because I was feeding a subRightSideInto() method call the
wrong assumptions.