Equals deduceInBool and defaultSimplification changes.
Don't be proactive about deduceInBool in Equals.__init__ -- that is costly and should not be necessary.  Fixed defaultSimplification by using isIrreducibleValue instead of checking if an object is a IrreducibleValue (the latter does not work on negated or multi-digit numbers).  Use TransRelupdater to be a little more efficient in applying desired transitivities rather than doing a transitivity seartch.