Adjustments w.r.t. specializing iterations.
Avoid going past the end of a tuple when getting its coordinates.  While this makes some logical sense, doing the start of each entry coordinate and one last one past the end of all entries, it forces requirements that should not be necessary (like '2+1=3' when only specializing with a 2-tuple.  Also change requirements, when specializing iterations, so they are of the form of differences being in Naturals rather than <= relations (so there are fewer types of requirements).