Changes w.r.t. multiple domain formatting of OperationOverInstances.
In particular, fixing errors when specifying multiple instance variables of a SetOfAll.  Also changing the way multiple domains are formatted -- as a list of membership conditions rather than the previous approach having a list of instance variables contained in a product of sets (this old way may be slightly more compact but is not so clear).