Begin to establish apply_permutation_thm in generic_methods.py
Largely based on the apply_commutation_thm() method in generic_methods.
Establishes the various tests for valid index inputs, the execution of the trivial (non) permutation, and the execution of the simple binary permutation.
Along the way, also included simple update to conclude() method in
SetEquiv class to allow the application of the concludeViaReflexivity() method form the SetEquiv class. Commit includes a temporary _testing_.ipynb in
the set_theory/enumeration context for testing the permutation work. That
nb can later be discarded.