Augment Set.deduceEnumSubsetEq() to work with subset input
The alternate Set.deduceEnumSubsetEqAlt() method now working well with
both subset_indices and subset arguments. Had to modify final step using
the subLeftSideInto() to just sub on the rhs of the innerExpr instead of the
Expr. Still just in the ALT version and not in the main version yet. Related Issues #77 , #69 , #62 .