Begin update of QPE/_common_, update SubIndexed class
The SubIndexed class is embedded in physics/quantum/QPE/phaseEstOps.py
and was modified to update to current Prove-It and allow some of the
QPE/_common_.ipynb content to be translated/updated. Not absolutely clear
how much of such translation is necessary, but it's a learning process. This commit includes an untouched copy of the original phaseEstOps.py.
imaginatively named phaseEstOps_old.py. Also had to temporarily comment
out some related __init__.py files until other components are updated.