Update Set.deduceEnumSubsetEq to now incorporate the Set() param
This commit temporarily relabels the old deduceEnumSubsetEq as
deduceEnumSubsetEqOld and relabels the deduceEnumSubsetEqAlt to be the
main deduceEnumSubsetEq method. Everything seems to be working fine.
Still to update the properSubset version so it too uses the Set input param.
This commit also cleans up some comments and notation in various places.
Related Issues: #96 , #77 , #71 , #69 .