Fixing expression notebook Literal/_operator_ bug(s).
Disabling automation to speed-up nonproof notebooks revealed a couple of issues w.r.t. expression notebooks that recreate Literal objects and _operator_ literals in particular.  First, 'context' information was being disregarded for 'remaking' Literal objects.  That was causing %check_expr checks to fail.  But this also revealed that operationClassOfOperator information was sometimes lacking.  This is improved by setting operationClassOfOperator elements within ExprType.__init__ rather than ExprType.__getattribute__ (do it as seen as the Operation class is instantiated instead of waiting for '_operator_' to be retrieved).