Allow proveit.reset() in script-mode.
That is, check of magics.proveItMagic exists before calling reset on it.  This enables resetting within a script.  This was done to experiment with turning off automation and resetting when loading theorems before recording proof dependencies in build.py, though it isn't clear that this is a good strategy.  So, it this feature isn't being used now, but may be useful at some point.