Finish updating quantum/QPE/_axioms_ (w/exceptions)
Finished updating quantum/QPE/_axioms_ with 3 exceptions that rely on the
Circuit class in the physics/quantum/ context. Unfortunately the Circuit class
still relies on the now-non-existent TensorExpression class. This commit also
cleans up some code comments and print() statements and introduces a
workaround for the latex \prescript command in the RegisterBra class in
quantumOps. Related Issues: #97 .