Add closure thms to exponentiation context, tweak Abs __init__
Abs __init__ calls the Operation superclass, but then was also trying to make
the additional assignment self.operand = param, but the self.operand was
already set byh the call to the Operation superclass. The closure thms added
to the exponentiation context are largely taken and adapted from the
_theorems_.old.py file in that same context and have some interesting
redundancies that still need cleaned up (also not yet finished transferring
all the old thms to the new _theorems_.ipynb. All in preparation for establishing
the deduceInNumberSet for the Exp class. Related Issues: #104 , #97.