Add fence=True in number/exponentiation/ Exp class formatted method
This appears to solve a grouping/fencing problem ( Issue #103 ) that was
cropping up in the Exp class when there was a Mult() nested within an Exp.
The fence=True was added to the args provided for the self.base.formatted()
method defining the inner_str. The change fixed some notational errors
appearing in exponentiation/_theorems_.ipynb. Related Issue: #97, #104.