Add natPosInRealsPos thm to number/sets/real context
And added this thm to the import statement at the end of the number/sets/reals/real.py, after which the system seemed to be able to now quickly prove that 2 \in RealsPos with no trouble where it was previously hanging/crashing. Commit also includes some cleaning and
line-length-reduction in code (particularly in the exp.py and particularly
various diagnostic/debugging print statements). The _demonstrations_.ipynb
still needs some general cleaning. Related Issues: #97, #104.