Fix TensorProd class call in quantum/_axioms_.ipnb
Also modified the latex portion of the literal operator definition for ScalarProd inside linalg/matrixOps.py. I had earlier tried to put a \; spacing marking but it was causing problems in constructing the expression; replacing that with \thinspace seems to have fixed this minor issue.
TensorProd class call was trying to use an explicit square-bracketed list
instead of a simple set of comma-separated args. This then cleared up the extractInitArgvalue NotImplementedError arising during a build.py --justessential.
Related issues: #97, #104.