Update two_pow_t_in_posnats to used deduceInNumberSet()
Now proves that 2^t is a natural positive number, using the
Exp.deduceInNumberSet() method and the axiomatic definition of t as a
positive natural number (t is denoted by the Literal t_ and represents the
number of Qbits in the input register). Commit also augments some axiom
and theorem cells with some explanatory comments.
Related Issues: #97, #104.