Update quantum, QPE, QFT contexts with automation=False
The initial import cells for the _axioms_, _common_, and _theorems_ in quantum, QPE, and QFT contexts augmented with the command
proveit.defaults.automation = False
to help speed things up and avoid the need for the try/except import of various
theorems in some .py files. Also cleans up proof that 2^t - 1 is in positive nats.
Related Issues: #97, #104.