Update real/interval.py for new theorems and context
This updates the real/interval.py so that calls to import real.theorems are
now replaced with calls to import ._theorems_ and also modifies several
theorem names to be consistent with recently updated theorems in
the number/sets/real context. Also updates and elaborates a bit on
the number/sets/real/_demonstrations_.ipynb page. Related Issues: #97, #116.