Update number/modular/_theorems_ to explicitly disallow modulus of 0
Disallowing modulus of 0 was achieved in the relevant theorems by including
a condition=[NotEquals(x, zero)] for the relevant modulus x. This commit also
fixes a formatting bug in the output of a SetOfAll expression, now suppressing
the empty condition |() when no explicit condition is provided in the SetOfAll
class call (achieved with a minor edit in the _formatted() method of the
SetOfAll class in logic/set_theory/comprehension/set_of_all.py. Commit also
includes some minor reformatting of set_of_all.py to make it more PEP8
compliant. Related Issues: #97, #113.