Started Indexed -> IndexedVar overhaul.
Renaming Indexed to IndexedVar.  Tentatively making it an Operation, with the Variable as the operator and the index as the operand.  The idea is then to use the machinery of operation substitution for it substitution.  Instead of multiple indices, allow nested IndexedVar's with a default style that flattens it as if it were multiple indices.  That will also simplify the substitution machinery.