"Instantiation" to replace "Specialization"
This new "instantiation" has the added feature that it can replace things inside and outside of the universal quantifiers.  For replacing things outside the universal quantifiers, they must be replaced on the "assumptions" side of the KnownTruth as well.  This can save an intermediate "generalize" step, but is still relatively simple and made to be compatible with new Iter substituted features and using Lambda.apply.