Change instanceVarOrVars to instanceParamOrParams, etc.
Distinguish the "parameters" from "parameter variables" in OperationOverInstanstances (affecting ForAll, Exists, NotExists, SetOfAll, Sum, etc.).  Just as for Lambda, a parameter may be an iteration over indexed variables.  The associated parameter variable would be the Variable being indexed.  So the "instanceParam" or "instanceParams" are for the parameters and the "instanceVar" or "instanceVars" for the parameter variables.