Simplified Iter expression structure (no Conditional lambda map)
Instead of having the Iter expression hide the start_index and end_index internal to a Conditional lambda map, just keep it simple by having the Iter have the basic lambda_map, start_index, and end_index as sub-expressions.  If we want to exploit the property that the parameter of an Iter is between the start and end indices, we could emply axioms/theorems explicitly and keep the complexity out of the core.