Modify Operation.string() and .latex() methods ...
to correct a tendency to omit important sub-fencing in outputs. This commit
also continues intermediate work on Round.doReducedSimplification() and
Round.roundingExtraction(). This marks a turning point where the dRS()
method can now apply a single iteration of roundingExtraction correctly to a
Round(Add(a, b, ..., n)) expression, finding the first suboperand that is already
an proven Integer that leaves each of the remaining suboperands as already-
proven Reals, and applying the roundingExtraction() method. Next on the
ToDo list is implementing an iterative version of this to repeatedly apply the
roundingExtraction() method until no longer possible. Related Issues: #97, #120.