Finish updating rounding_methods.py
and finish converting the doReducedSimplification(), roundingElimination(), roundingExtraction(), and deduceInNumberSet() methods in the Ceil, Floor, and Round classes to utilize the more general rounding_methods.py functions. This provides some good consolidation of the code and still maintains the equivalence-method status of the roundingElimination() and roundingExtraction() methods. This commit also includes some continued reogranizing of the testing section of the rounding/_demonstrations_.ipynb and some other general code clean-up. Related Issues: #97, #120.