Core updates in getting an induction proof (mult_substitution) to work
This required resolving some tricky issues with respect to single vs multiple (including with ExprRange entries) assumptions, operands, and conditions.  Introduced automatic collapse of an ExprRange with the same start and end and dealt with making the generic version of a Lambda with IndexedVar parameters (this arrises when ExprRanges collaps).  Also some InnerExpr fixes to deal with substitutions of multiple operands.  Preparing InnerExpr for more capabilities, relabeling of KnownTruth assumptions, but that isn't finished.