Augment rounding_methods.apply_reducedSimplification() fxn
The augmentation is in two parts: (1) adds the InSet.knownMemberships() method check as a second step after finding that the rounding f(x) operand x is not immediately known to be an integer; (2) to that new step, and to the following step where the InSet.knownMemberships() is used to check sub-operands, added the check that the KnownTruth.isSufficient(assumptions) returns True before also checking if the known truth involves a set that is a subset of our desired set (typically our desired set is Integers). This all seems to be working well, but code still needs cleaned up now and numerous diagnostic print statements removed. Related Issues: #97, #120.