Establish new dictionaries in Equals class.
The new dictionaries are known_simplifications and known_evaluation_sets, each of which use the tuple (knownTruth.lhs, tuple(sorted(assumptions))) as a key and the knownTruth as a the value. These dictionaries are not used for anything yet but diagnostic print statements suggest they are storing the information as intended (for later use as substitutes for the simplifications and evaluations dictionaries, allowing more precise use based on the more detailed keys). The Equals class code and the rounding_methods.py code is currently littered with testing print statements. Related Issues: #97, #120.