Core updates: derivation rule protection, auto-relabeling, etc.
Expression methods from being overridden so we can trust that the derivation rules are not thwarted be code external to the core.  For the same "trust" reason, use "_checked_make" to check that an expression is remade properly (since "_make" can be influenced external to the core).  Automatically relabel parameters to avoid collisions instead of raising a ScopingViolation error (ScopingViolation was entirely removed).  With this change, "reseved_vars" are no longer needed in the "_substituted" methods.  Made Conditional treat only a single value/condition pair for simplicity and added ConditionalSet for the multi-condition case.  Worked on some reduction rules (for Conditions, Operations, and ExprRules): established "defaults" flags to enable/disable them and adding appropriate requirements when they are employed.  Various other core fixes.