Update Equals class with full impl of known_simplifications dictionary
This seems to now be a working version in which the original Equals.simplifications dictionary in the Equals class code has been fully replaced by the Equals.known_simplifications dictionary which differs from the original in having just a single, most-recent entry for each key expression instead of a list of possible simplifications for each key expression. The code is still a mess, littered with print statements and some testing still needs to be done, but looks quite promising. Related Issues: #81, #97, #120.