Clean up code
This commit includes an odd file abs_temp.py to save away some changes in the modular/abs.py that I believe were inadvertently made in this branch but were intended for a different branch. Will check on this in the near future. Reverting the changes in the modular/abs.py seems to cause no issues in this branch. Somewhat puzzling — could simply be I got confused about which branch I was working in.