Establish Abs.doReducedSimplification()
This covers the simple cases for Abs(x) where x is already known or assumed to be a non-negative real (returning |– Abs(x) = x), known or assumed to be a negative real (returning |– Abs(x) = -x, known or assumed to be in a subset of positive reals, and (more iffy right now) -x is known or assumed to be in a subset of positive reals which means that x would itself be in negative reals. Along the way we also: (1) slightly modified the Neg.doubleNegSimplification() method, replacing the simple assumptions in the specialize call to assumptions=assumptions (which helped deal with an error arising occasionally); (2) added a thm to number/sets/real context claiming that x in pos reals reals means -x is in negative reals. Related Issues: #97, #119.