various core fixes/changes: 'equiv_alt_expansions' etc.
Fixes w.r.t. recording Instantiation mappings and allowing versatility in using implicit instantiation of a range in addition to equivalent alternative expansions.  Fixes related to partial masking (of variable ranges) and relabeling inner expressions (dealing with OperationOverInstances properly).  Provide assumptions to an InnerExpr object when it is created (necessary when dealing with slices with ExprRanges).  ExprTuple.merger fixes.  A change to the formatted of ExprRanges when using explicit parameterization to be more compact and clear: x_1, ..x_k.., x_n