core updates: free_var_ranges mods, generalize fixes, auto_reduce, etc.
Don't auto_reduce when making generic version of lambda expression.  Simplified 'free_var_ranges' to only eliminate explicitly-covered ranges (no longer eliminating ranges with internally bound variables which was probably motivated by erroneous implementations of 'instantiate'.  Allow 'generalize' to accept more forms of forall variables (singular, lists or lists of lists).  Turn off 'allow_relabeling' for instantiating assumptions.  Fixes related to the 'instantiation' mappings that are reported.  'globalRepl' returns a multi-parameter lambda map when replacing an ExprTuple of operands.  The 'conditions' of an OperationOverInstances will be the "And" version when singular to distinguish it from not having "And" at all.