Remove chain-specializations from interval.py code
This reduces and simplifies the code in interval.py, rewriting the double-specialization calls for all the specializations of theorems involving nested Foralls (i.e. instead of having thm.specialize({a:x}).specialize({b:y}), we just use thm.specialize({a:x, b:y})). I suspect such code was left over from a time when we had no choice but to call the specializations separately for each of the nested Foralls. Related Issue: #116.