Update number/sets/real/_demonstrations_.ipynb with 2nd demo
The 2nd demo goes through steps to prove that [1,2] is an improper subset of the positive reals. Turns out that it considerably more difficult to show that with a strict/proper subset instead. Commit includes a minor aesthetic edit to the number/sets/real/_theorems_.ipynb as well. Related Issue: #116.