Add 8 interval subset thms to number/sets/real context
The 8 thms correspond to the 4 possible open/closed interval combinations, combined with the two possible number sets RealsPos and RealsNonNeg for a collection of thms of the form: Forall a, b in SET, the interval (a,b) is a subset of SET. For example, the interval (a, b) is a proper subset of the RealsPos whenever a,b are positive reals. Related Issue: #116.