reduceOperands in logic.equality.equals.py must skip ExprRange operands
ExprRange operands may not be reduced in the same way other operands can be.  We don't have a generic way to deal with this, for now.  The And class, for example, has its own way of dealing with ExprRange operands that may be simplified.  For now, we'll do this individualized approach to ExprRange operands.  Maybe we will come up with a more generic, reusable solution in the future.