Resetting to eliminate CONTRIBUTING and sqrt(2)-related changes
This commit then appears to include a recent merge of master into this branch as well as numerous changes previously introduced in separate commits in developeing the number/real/absolute_value context. This makes the branch content somewhat more “pure” but also compresses the commit history in a way that might be problematic later. The CONTRIBUTING-related docs and the sqrt(2)-not-rational related docs should now safely reside in other branches. Related Issues: #93, #97, #119, #139.