Reorganize and finish up the proof of t4_alt
Eliminates a number of redundant and/or obsolete cells and reorganizes with more step-by-step headers and a bit of transitional text. Also adds the final %qed, which produces the formal proof output (interestingly accepting the presence of the b' assumption, which is not present in the statement of the theorem at the top of the theorem proof page). Related Issue: #93.