Establish in_IntervalXX_def axioms for number/sets/real context
These axioms define what it means to have x in a real interval — for example, x in the real interval (a, b) means that x is real and a < x < b. The axioms will eventually be useful when establishing the membershipObject() method for each Interval type and the related IntervalMembership class. Related Issue: #116.