Core updates: tracking "equality replacement requirements"
Track "equality replacement requirements" that are employed during lambda application (or instantiation) via "auto_reduced" methods.  These are marked and displayed with "askerisks" in the proof so it is clear to any proof checker what auto-reductions are being applied during an instantiation.  Also remove '_simplifiedCoord' which is no longer used, and fixed a bug in Operation._replaced (where 'replaced' was called when it should call '_auto_reduced').