Update logic/set_theory/enumeration/_testing_.ipynb to correct errors
Cleaning up some small errors in this testing nb in preparation for some more general changes and consolidations. This commit simply corrects the name for a small enumerated set and corrects the importation of the obsolete reduction theorem to the importation of the reduction_left and reduction_right theorems and implements a brief specialization testing of those theorems. Related Issue: #69.