Add inEnumeratedSet theorem to set_theory/enumeration context
The inEnumeratedSet thm basically says that B is an element of {A, B, C}, and might serve as a more convenient and efficient definition of enumerated set membership than the axiomatic def involving disjunction. Added this theorem because of the rather long proofs being created in Prove-It for simple membership claims like 1 \in {1, 2, 3}. Related Issue: #69.