Core updates: free_vars with 'err_inclusively', possibly_free_var_ranges
The free_vars function now requires an 'err_inclusively' flag.  When true, it may report more free variables than actual if there is ambiguity with respect to index ranges.  When false, it may report fewer free variables than actual.  For example,
   (x_1, ..., x_n) -> x_1 + ... + x_k + x_{k+1} + ... + x_{n}
will report x only if err_inclusively=True.  Also,
   (x_1, ..., x_n) -> x_1 + ... + x_n + x_{n+1}
will not report x if err_inclusively=False.
The strictness of these bounds may be tightened in the future (e.g., accounting for literal integer offsets) but there will always be ambiguous cases for a given expression in the absense of assumptions.  free_var_ranges was changed to possibly_free_var_ranges because it always errors inclusively.  Also fixed an issue with 'traverse_inner_expressions' -- if it is given a KnownTruth, report the expression.  This was caught because it caused "Lambda.globalRepl" to do the wrong thing.