Update enumeration/_demos_.ipynb to test deduceEnumSubsetEq method
The deduceEnumSubsetEq() method is working as designed, but needs some redesign to make sure it does not treat an enumerated Set as a possible multi-set. Currently the deduceEnumSubsetEq() method interprets a set such as {a, b, a} as a multi-set and the set {a, b, b, a} as NOT a subset of {a, b, a} because of the multiplicity of the b element. This is feature/bug left over from my original coding some months back when I was interpreting a Set as a multi-set. Related Issue: #69.