Finish deduceEnumSubsetEq() and clean up enum_set.py
This finishes the Set.deduceEnumSubsetEq() method, which now seems to work well and treats multiplicities in a Set as it should (i.e. a Set with multiplicities only appears to be a multi-set but actually is not and is equal to its support set). Also cleans up the Set class more generally, removing a number of now obsolete code and comments. Related Issue: #69.