KnownTruth updates: 'num_forall_eliminations' and keyword only parameters.
KnownTruth.specialize now takes an optional 'num_forall_eliminations' keyword only parameter which is useful when you want to relabel the parameters of an inner nested forall rather than eliminating the forall.  'assumptions' is also now a keyword only parameter.  Also, KnownTruth.__getattr__ is updated to properly deal with methods with keyword only parameter and 'assumptions' as a keyword only argument in particular.  Either way (keyword-only or not) the KnownTruth assumptions will be automatically included.