Add equalElementEquality thm to set_theory/enumeration context
The equalElementEquality theorem basically says that an enumerated set such as {a, b, c} must be equal to the enumerated set {a, d, c} when b = c. This is being introduced to allow such equalities to be deduced in some of the machinery for the Set.deduceEnumSubsetEq() and Set.deduceEnumProperSubset() methods. Related Issues: #69, #77.