Establish 1st draft of Set.single_elem_substitution() method
The Set.single_elem_substitution() method deduces an equality between a given enumerated Set and the same eumerated set when a equal substitution is made for a single element. For example, when x = 2, it will deduce that |– {x, y, z}={2, y, z}. This first draft simply implements a single argument and specializes the related equalElementEquality theorem. Related Issues: #69, #77.