Finish Set.single_elem_substitution() method in enumeration context
This completes the Set.single_elem_substitution() method in enumeration context, allowing us to deduce that an enumerated set is equal to the set obtained when a specific element is replaced by something equal to that element. Hope now to use this in a more general Set.elem_substitution() method, which in turn will be used in the Set.deduceEnumSubsetEq() method. Related Issues: #69, #77.