Establish the Set.elem_substitution() method in enumeration context
The Set.elem_substitution() method deduces an equality between a given enumerated Set and the Set obtained when all instances of a specified element are replaced by a substitution value. For example, when x = 2, it will deduce that |– {x, y, x, 3, z}={2, y, 2, 3, z}. This is intended to be used in the Set.deduceEnumSubsetEq() method and depends on the earlier-established Sset.single_elem_substitution() method. Related Issues: #69, #77.