Various core, logic, and number updates to fix demonstrations.
Added Expression.disproven,  Fixed Expression.copy.  Evaluation/simplification updates: automation as a keyword argument and **kward passed through in case we want to pass other directives to doReducedEvaluation or doReducedSimplification.  deduceInNumberSet should always accept 'assumptions' (even if they are needed).  Various proveit.logic and proveit.number updates to get demonstrations to pass.  And.evaluation and Or.evaluation overhauled to be efficient (i.e., first check if any operand is known to evaluate to True, for Or, or False, for And.  And.unaryReducion and Or.unaryReduction update.  Forall.conclude update to try a 'generalization with automation' strategy if there is no 'foldAsForall' to try (staying consistent with philosophy of only one nontrivial conclude strategy).  InSet.conclude and NotInSet.conclude updates to be consistent with the "one nontrivial strategy" rule of thumb.  Add and Mult theorem-application method updates.