Upgrade Set.deduceEnumProperSubset() method
Upgrade the Set.deduceEnumProperSubset() method to make use of the recently created Set.elem_substitution() method, allowing us to deduce things like {x=2} |– {x, y, z} \subset {2, y, z, f}. The upgrade appears to be complete and working well, but some code editing still in progress. This commit also performs the substitution and reduction of the subset argument earlier than in the deduceEnumSubsetEq() method, suggesting we should go back to that and rearrange a bit as well. Related Issues: #69, #77.