Correct method name typo in set_theory/membership/in_set.py
The deriveSupersetMembership() method was misspelled. Actually: this was misspelled in the set_theory/containment/subset.py file and was somehow carried over to be used with that misspelling in in_set.py code. Rather strange actually. So when this was corrected in the subset.py code, had to update the in_set.py code as well.