Upgrade applyTransitivity() meth in subset classes to handle equalities
The equality conditions for the applyTransitivity() methods in the Subset, ProperSubset, and SubsetEq classes all tried to call a nonexistent ContainmentRelation.applyTransitivity() method. Later might still try to augment the ContainmentRelation class to deal with this but now the Equals condition handles it within the subset classes themselves with a subXSideInto() approach. Related Issue: #69.