Fix SubsetEq.unfold() meth to conform to new forall.instantiate()
The old SubsetEq.unfold() method sent along an elemInstanceVar kwarg to the new instantiate() method with the intent of using it to replace the instance var using a relabelMap, but the relabelMap has been eliminated. This solution still allows the user to specify the eleminstanceVar to be relabeled but uses an extra instantiation step to achieve using the num_forall_eliminations=0 kwarg. Related Issue: #69.