Update/fix logic/boolean/disjunction Or.unaryReduction() method calls
Not absolutely clear what is happening here, but it appears that the Or.unaryReduction() method was eliminated or changed to Or.deduceUnaryEquiv, and the deduceUnaryEquiv itself was importing the wrong theorem from the logic/boolean/disjunction context. Instead of wholesale replacing calls to the unaryReduction() method and simply deleting lines in the deduceUnaryEquiv() method, I've commented out the offending lines and written replacements (just in case I've misinterpreted here). Related Issue: #69.