Reestablish the Or.unaryReduction() call in disj/_demos_.ipynb
In further testing and evaluation, discovered that most consistent with recent merge of master into craft_edits_containment_demos branch is actually the opposite of what I originally thought: need to keep the Or.unaryReduction() method instead of the deduceUnaryEquiv() method. This commit re-establishes the use of the unaryReduction() method example in the disjunction/_demonstrations_.ipynb. A separate commit will clean up the related method definition and related calls in the or_op.py file. Related Issue: #69.