Rename literal bPrime in root2…/_theorems_.ipynb to resolve conflict
In developing some sub-proofs for some of the supporting theorems used in the proof of sqrt(2) not rational, inadvertently redefined bPrime as a literal when bPrime was also needed later in the notebook as a variable, so this commit redefines the literal bPrime as a bPrimePrime (and in a separate commit I've reworked one of the sub-theorem proofs to use the bPrimePrime literal as needed). Related Issue: #93.