Update proof to use b'' instead of b', avoiding conflict with b' variable
In developing some sub-proofs for some of the supporting theorems used in the proof of sqrt(2) not rational, inadvertently redefined bPrime as a literal when bPrime was also needed later in the notebook as a variable, so an earlier commit redefines the literal bPrime as a bPrimePrime and this commit makes the substitution(s) as needed throughout the associated proof notebook. Related Issue: #93.