Update sqrt(2) not rational-related material for recent merge
Update sqrt(2) not rational-related material for recent merge of master in branch 93. This consists primarily of replacing a number of specialize() calls with instantiate() calls (which is not strictly always necessary, but simplifies matters and corrects some issues when arguments are omitted), corrects a few cases where kwargs were renamed (domainLists have become domain_lists in generalize() method), and renames a literal bPrime to bPrimePrime to avoid conflicts with a variable with the name bPrime.